Refactor AuthRequest Logic into Service
