+++ src/Api/Auth/Controllers/AuthRequestsController.cs
@@ -1,14 +1,11 @@
-﻿using Bit.Api.Auth.Models.Request;
-using Bit.Api.Auth.Models.Response;
+﻿using Bit.Api.Auth.Models.Response;
 using Bit.Api.Models.Response;
-using Bit.Core.Auth.Entities;
-using Bit.Core.Auth.Exceptions;
-using Bit.Core.Context;
+using Bit.Core.Auth.Models.Api.Request.AuthRequest;
+using Bit.Core.Auth.Services;
 using Bit.Core.Exceptions;
 using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Bit.Core.Settings;
-using Bit.Core.Utilities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -18,30 +15,21 @@ namespace Bit.Api.Auth.Controllers;
 [Authorize("Application")]
 public class AuthRequestsController : Controller
 {
-    private readonly IUserRepository _userRepository;
-    private readonly IDeviceRepository _deviceRepository;
     private readonly IUserService _userService;
     private readonly IAuthRequestRepository _authRequestRepository;
-    private readonly ICurrentContext _currentContext;
-    private readonly IPushNotificationService _pushNotificationService;
     private readonly IGlobalSettings _globalSettings;
+    private readonly IAuthRequestService _authRequestService;
 
     public AuthRequestsController(
-        IUserRepository userRepository,
-        IDeviceRepository deviceRepository,
         IUserService userService,
         IAuthRequestRepository authRequestRepository,
-        ICurrentContext currentContext,
-        IPushNotificationService pushNotificationService,
-        IGlobalSettings globalSettings)
+        IGlobalSettings globalSettings,
+        IAuthRequestService authRequestService)
     {
-        _userRepository = userRepository;
-        _deviceRepository = deviceRepository;
         _userService = userService;
         _authRequestRepository = authRequestRepository;
-        _currentContext = currentContext;
-        _pushNotificationService = pushNotificationService;
         _globalSettings = globalSettings;
+        _authRequestService = authRequestService;
     }
 
     [HttpGet("")]
@@ -54,11 +42,12 @@ public class AuthRequestsController : Controller
     }
 
     [HttpGet("{id}")]
-    public async Task<AuthRequestResponseModel> Get(string id)
+    public async Task<AuthRequestResponseModel> Get(Guid id)
     {
         var userId = _userService.GetProperUserId(User).Value;
-        var authRequest = await _authRequestRepository.GetByIdAsync(new Guid(id));
-        if (authRequest == null || authRequest.UserId != userId)
+        var authRequest = await _authRequestService.GetAuthRequestAsync(id, userId);
+
+        if (authRequest == null)
         {
             throw new NotFoundException();
         }
@@ -68,10 +57,11 @@ public class AuthRequestsController : Controller
 
     [HttpGet("{id}/response")]
     [AllowAnonymous]
-    public async Task<AuthRequestResponseModel> GetResponse(string id, [FromQuery] string code)
+    public async Task<AuthRequestResponseModel> GetResponse(Guid id, [FromQuery] string code)
     {
-        var authRequest = await _authRequestRepository.GetByIdAsync(new Guid(id));
-        if (authRequest == null || !CoreHelpers.FixedTimeEquals(authRequest.AccessCode, code) || authRequest.GetExpirationDate() < DateTime.UtcNow)
+        var authRequest = await _authRequestService.GetValidatedAuthRequestAsync(id, code);
+
+        if (authRequest == null)
         {
             throw new NotFoundException();
         }
@@ -83,80 +73,16 @@ public class AuthRequestsController : Controller
     [AllowAnonymous]
     public async Task<AuthRequestResponseModel> Post([FromBody] AuthRequestCreateRequestModel model)
     {
-        var user = await _userRepository.GetByEmailAsync(model.Email);
-        if (user == null)
-        {
-            throw new NotFoundException();
-        }
-        if (!_currentContext.DeviceType.HasValue)
-        {
-            throw new BadRequestException("Device type not provided.");
-        }
-        if (_globalSettings.PasswordlessAuth.KnownDevicesOnly)
-        {
-            var devices = await _deviceRepository.GetManyByUserIdAsync(user.Id);
-            if (devices == null || !devices.Any(d => d.Identifier == model.DeviceIdentifier))
-            {
-                throw new BadRequestException("Login with device is only available on devices that have been previously logged in.");
-            }
-        }
-
-        var authRequest = new AuthRequest
-        {
-            RequestDeviceIdentifier = model.DeviceIdentifier,
-            RequestDeviceType = _currentContext.DeviceType.Value,
-            RequestIpAddress = _currentContext.IpAddress,
-            AccessCode = model.AccessCode,
-            PublicKey = model.PublicKey,
-            UserId = user.Id,
-            Type = model.Type.Value
-        };
-        authRequest = await _authRequestRepository.CreateAsync(authRequest);
-        await _pushNotificationService.PushAuthRequestAsync(authRequest);
+        var authRequest = await _authRequestService.CreateAuthRequestAsync(model);
         var r = new AuthRequestResponseModel(authRequest, _globalSettings.BaseServiceUri.Vault);
         return r;
     }
 
     [HttpPut("{id}")]
-    public async Task<AuthRequestResponseModel> Put(string id, [FromBody] AuthRequestUpdateRequestModel model)
+    public async Task<AuthRequestResponseModel> Put(Guid id, [FromBody] AuthRequestUpdateRequestModel model)
     {
         var userId = _userService.GetProperUserId(User).Value;
-        var authRequest = await _authRequestRepository.GetByIdAsync(new Guid(id));
-        if (authRequest == null || authRequest.UserId != userId || authRequest.GetExpirationDate() < DateTime.UtcNow)
-        {
-            throw new NotFoundException();
-        }
-
-        if (authRequest.Approved is not null)
-        {
-            throw new DuplicateAuthRequestException();
-        }
-
-        var device = await _deviceRepository.GetByIdentifierAsync(model.DeviceIdentifier, userId);
-        if (device == null)
-        {
-            throw new BadRequestException("Invalid device.");
-        }
-
-        authRequest.ResponseDeviceId = device.Id;
-        authRequest.ResponseDate = DateTime.UtcNow;
-        authRequest.Approved = model.RequestApproved;
-
-        if (model.RequestApproved)
-        {
-            authRequest.Key = model.Key;
-            authRequest.MasterPasswordHash = model.MasterPasswordHash;
-        }
-
-        await _authRequestRepository.ReplaceAsync(authRequest);
-
-        // We only want to send an approval notification if the request is approved (or null), 
-        // to not leak that it was denied to the originating client if it was originated by a malicious actor.
-        if (authRequest.Approved ?? true)
-        {
-            await _pushNotificationService.PushAuthRequestResponseAsync(authRequest);
-        }
-
+        var authRequest = await _authRequestService.UpdateAuthRequestAsync(id, userId, model);
         return new AuthRequestResponseModel(authRequest, _globalSettings.BaseServiceUri.Vault);
     }
 }
