+++ src/Core/Auth/Models/Api/Request/AuthRequest/AuthRequestCreateRequestModel.cs
@@ -1,8 +1,7 @@
-﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations;
 using Bit.Core.Auth.Enums;
-using Newtonsoft.Json;
 
-namespace Bit.Api.Auth.Models.Request;
+namespace Bit.Core.Auth.Models.Api.Request.AuthRequest;
 
 public class AuthRequestCreateRequestModel
 {
@@ -17,14 +16,4 @@ public class AuthRequestCreateRequestModel
     public string AccessCode { get; set; }
     [Required]
     public AuthRequestType? Type { get; set; }
-}
-
-public class AuthRequestUpdateRequestModel
-{
-    public string Key { get; set; }
-    public string MasterPasswordHash { get; set; }
-    [Required]
-    public string DeviceIdentifier { get; set; }
-    [Required]
-    public bool RequestApproved { get; set; }
 }
